Address review feedback

+++ libcaf_net/caf/net/http/client.cpp
@@ -84,8 +84,7 @@ public:
 
   void request_messages() override {
     if (!down_->is_reading())
-      down_->configure_read(
-        receive_policy::up_to(static_cast<uint32_t>(max_response_size_)));
+      down_->configure_read(receive_policy::up_to(max_response_size_));
   }
 
   void suspend_reading() override {
